+++ packages/keyring-service/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @unisat/keyring-service
 
+## 0.1.25
+
+### Patch Changes
+
+- feat: add readonly keyring
+
 ## 0.1.24
 
 ### Patch Changes