+++ packages/tx-helpers/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @unisat/tx-helpers
 
+## 0.1.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @unisat/keyring-service@0.1.25
+
 ## 0.1.10
 
 ### Patch Changes